Upgrade iron-router to 0.8.1. #32
This needs a little tweaking. You need a recent version of npm, and upgrading npm with sudo threw permission errors on my system. I ran something along the lines of:

    sudo npm update -g npm
    sudo find ~/.npm -user root -exec chown USER.USER '{}' \;

That reset the permissions errors in my `~/.npm` folder. Then I was able to restart `mrt` and it was able to upgrade the dependencies.